refs/heads/main: Guild new panels

+++ Assets/Scripts/Network/DataServer/Models/Guilds/RecruitData.cs
@@ -8,7 +8,7 @@
         public float DonateMantissa { get; set; }
         public int DonateE10 { get; set; }
         public float ResultMantissa { get; set; }
-        public float ResultE10 { get; set; }
+        public int ResultE10 { get; set; }
         public string IntroductionDateTime { get; set; } = string.Empty;
         public int CountRaidBoss { get; set; }
         public string DateTimeFirstRaidBoss { get; set; } = string.Empty;
